requirements.txt
schedule
pyyaml
pyttsx3
selenium
webdriver-manager
imapclient
flask
openai
requests
pyyaml


